[Feat]학식 데이터 캐싱 완료
